--- a/idl/gossip_digest.idl.hh
+++ b/idl/gossip_digest.idl.hh
@@ -74,4 +74,8 @@ class gossip_digest_ack2 {
     std::map<gms::inet_address, gms::endpoint_state> get_endpoint_state_map();
 };
 
+class cluster_ring_status_response {
+    std::map<sstring, sstring> status;
+};
+
 }