@asias 2019-10-09
Depends on:  "sharded: Make map work with mapper that returns a future" seastar patch.

Updates since last version:
- Hash pending ranges as well
- Use key value for GOSSIP_QUERY_TOKEN_STATUS response so that we can extend the protocol later if we want to checksum more data or change the hash method
- Rebase to latest master

Previous updates:
- respect skip_wait_for_gossip_to_settle
- no check replace_address again inside replacing_a_node_with_diff_ip
- log exception from gossiper::wait_for_gossip_with_query
- hash both TOKENS and STATUS from gossip