@asias 2019-10-09
Hello guys,  let's try to make some progress. This patch set is on the list for almost 3 months since the previous PR and almost 10 months since the first RFC submission. It is a not very productive to switch back and forth.

@tgrabiec I believe all the previous comments are addressed. This version makes sure all the shards and not just shard zero see the same view.