@slivne 2019-11-19
@avikivity  / @tgrabiec - we have this patchset that reduces the overall bootstrap time - it will help a lot in our testing (and not only in testing), can we make progress with this
 - AFAIK all the comments that have been raised have been addressed 
 - is there something I/ Asias are missing ?
 - is there some other concern we need to handle ?