@tgrabiec 2019-12-04
On Tue, Nov 19, 2019 at 10:24 AM Shlomi Livne <notifications@github.com>
wrote:

> @avikivity <https://github.com/avikivity> / @tgrabiec
> <https://github.com/tgrabiec> - we have this patchset that reduces the
> overall bootstrap time - it will help a lot in our testing (and not only in
> testing), can we make progress with this
>
>    - AFAIK all the comments that have been raised have been addressed
>
>
No

>
>    - is there something I/ Asias are missing ?
>
>
Yes


>
>    - is there some other concern we need to handle ?
>
>