@slivne 2019-12-04
@asias can you try and address what tomek is missing