@asias 2019-12-04
On Wed, Dec 4, 2019, 20:55 Shlomi Livne <notifications@github.com> wrote:

> @asias <https://github.com/asias> can you try and address what tomek is
> missing
>
I wish I could. But how are we supposed to know what he think is missing
from a single word review?

>
>