@asias 2019-12-10
Here is a combined change log. 

    Make sure all shards has the same view
    Hash pending ranges as well
    Use key value for GOSSIP_QUERY_TOKEN_STATUS response so that we can extend the protocol later if we want to checksum more data or change the hash method
    Respect skip_wait_for_gossip_to_settle
    No check replace_address again inside replacing_a_node_with_diff_ip
    Log exception from gossiper::wait_for_gossip_with_query
    Hash both TOKENS and STATUS from gossip
    Rebase to latest master

With this series, we are much much safer. We now even can make sure all the shards have the same view before the node proceeds. Another nice thing is that the change is not intrusive, we can switch to other solutions later very easily, only changing the procedure of the booting node is enough.