@slivne 2019-12-17
@tgrabiec can you please help us find the missing item / unaddressed comment(s) - so we can move forward