@asias 2020-03-12
Changes in 49b14a914db83bc1755fd88a6b8165b017bd0499

- Rebased to latest master

@avikivity As we discussed in the thread "[PATCH v1] gossip: make sleeping intervals configurable", you seem to have positive feedback. Can you please review?