@asias 2020-03-17
> This PR does two separate things:
> 
>     * one is fixing node replacement behavior
> 
>     * the other is what the PR title actually says, making boot process faster.
> 
> 
> I think it should be split into separate PRs.

The change to the replace is a requirement for boot faster to work properly. I have no problem make it a separate PR. At this stage, I would like to put them into one to reduce the inter-PR dependencies.