@asias 2020-03-19
Changes in 9da855472c8afbd54d000ea5fee9c4223cc36da9:
- Rename verb from GOSSIP_QUERY_TOKEN_STATUS to CLUSTER_RING_STATUS
- Decouple the algorithm from gossiper class. They are now static function in gossiper.cc. 
- Add more comments